Ensure workspace shell renders for missing boards and channels
