+++ app/[boardId]/default.tsx
@@ -1,4 +1,4 @@
-import { BoardWorkspaceNotFound } from "./not-found";
+import { BoardWorkspaceNotFound } from "./board-workspace-not-found";
 
 export default function BoardDefaultSlot() {
   return <BoardWorkspaceNotFound />;
