+++ app/not-found.tsx
@@ -1,0 +1,5 @@
+import { BoardWorkspaceNotFound } from "./[boardId]/board-workspace-not-found";
+
+export default function NotFound() {
+  return <BoardWorkspaceNotFound />;
+}
