+++ pages/404.tsx
@@ -1,0 +1,5 @@
+import { BoardWorkspaceNotFound } from "../app/[boardId]/board-workspace-not-found";
+
+export default function NotFoundPage() {
+  return <BoardWorkspaceNotFound />;
+}
